A bit of logging, stats for ServiceJobRunner; moved to separate file; keep reference in ServiceFacadeImpl to check status, etc
